Update patch set 1
Patch Set 1:

I recall reading in #apps that you recently tried to update the webview to v62 but found the theme color patch did not apply. The change in Chromium which broke the patch was not that involved, so I updated the patch. With this the patch applies and builds.

I'm not sure how to test that the feature still works, however. Is there a manual test case somewhere to verify the extension still works as expected?

Patch-set: 1
Reviewer: Branden Archer <18383@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0